% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spps_basic_functions.R
\name{dexpacotm}
\alias{dexpacotm}
\title{This function computes the derivative of pi(x) for the SPPS model}
\usage{
dexpacotm(x, epsilon, delta)
}
\arguments{
\item{x}{A real number}

\item{epsilon}{A real number between 0 and 1}

\item{delta}{A real number between 0 and 1}
}
\value{
A real number
}
\description{
This function computes the derivative of pi(x) for the SPPS model
}
