% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spps_basic_functions.R
\name{dexpit}
\alias{dexpit}
\title{This function computes the derivatibe of the expit function}
\usage{
dexpit(x)
}
\arguments{
\item{x}{A real number}
}
\value{
A real number
}
\description{
This function computes the derivatibe of the expit function
}
