% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubly_protected_estimators.R
\name{dp}
\alias{dp}
\title{computes a doubly protected estimator of the mean of a random variable Y with the robust method based on maximum likelihood methods}
\usage{
dp(x, y, ps = NULL)
}
\arguments{
\item{x}{A matrix of covariates without an intercept (either \code{x} or \code{ps} muste be provided)}

\item{y}{A vector of outcomes (may contain NAs)}

\item{ps}{The estimated propensity score, i.e. the probability of y being observed conditional to \code{x} (either \code{x} or \code{ps} muste be provided)}
}
\value{
The estimated mean of Y
}
\description{
computes a doubly protected estimator of the mean of a random variable Y with the robust method based on maximum likelihood methods
}
\examples{
  vars <- rsamplemissing(n, alpha, beta)
  x <- vars$x
  y <- vars$y
  a <- vars$a
  dp(x, y)
}
