% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubly_protected_estimators.R
\name{drdpn}
\alias{drdpn}
\title{Computes a robust doubly protected estimator of a quantile of a random variable Y with the method proposed in Sued, Valdora and Yohai (2019) and the SPPS model for the propensity score.}
\usage{
drdpn(x, y, ps = NULL, type = 2, qq = 0.5)
}
\arguments{
\item{x}{A matrix of covariates without an intercept}

\item{y}{A vector of outcomes (may contain NAs)}

\item{type}{1 or 2; if type=2 the estimator is normalized as explained in Sued, Valdora and Yohai (2019)}
}
\value{
The estimated quantile of Y
}
\description{
Computes a robust doubly protected estimator of a quantile of a random variable Y with the method proposed in Sued, Valdora and Yohai (2019) and the SPPS model for the propensity score.
}
\examples{
  vars <- rsamplemissing(n, alpha, beta)
  x <- vars$x
  y <- vars$y
  drdpn(x, y)
}
