% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spps_basic_functions.R
\name{expacotm}
\alias{expacotm}
\title{This function computes \pi(x) for the SPPS model}
\usage{
expacotm(x, epsilon, delta)
}
\arguments{
\item{x}{A real number x}

\item{epsilon}{A real number between 0 and 1}

\item{delta}{A real number between 0 and 1}
}
\value{
A real number
}
\description{
This function computes \pi(x) for the SPPS model
}
