% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spps_basic_functions.R
\name{expacotminv}
\alias{expacotminv}
\title{Computes the inverse of the function pi(x) for  the spps  model.}
\usage{
expacotminv(y, epsilon, delta)
}
\arguments{
\item{y}{A real number between 0 and 1}

\item{delta}{A real number between 0 and 1}

\item{espilon}{A real number between 0 and 1 such that \code{epsilon} + \code{delta} < 1}
}
\value{
A real number
}
\description{
Computes the inverse of the function pi(x) for  the spps  model.
}
