% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_basic_functions.R, R/spps_basic_functions.R
\name{expit}
\alias{expit}
\title{This function computes exp(x) / (1 + exp(x)), the inverse of the logistic function log(p / (1 - p))}
\usage{
expit(x)

expit(x)
}
\arguments{
\item{x}{A real number}

\item{x}{A real number}
}
\value{
A real number between 0 and 1

A real number between 0 and 1
}
\description{
This function computes exp(x) / (1 + exp(x)), the inverse of the logistic function log(p / (1 - p))

This function computes exp(x) / (1 + exp(x)), the inverse of the logistic function log(p / (1 - p))
}
