% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spps_basic_functions.R
\name{expitinv}
\alias{expitinv}
\title{This function computes  the logistic function log(p / (1 - p))}
\usage{
expitinv(p)
}
\arguments{
\item{p}{A real number between 0 and 1}
}
\value{
A real number
}
\description{
This function computes  the logistic function log(p / (1 - p))
}
