% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw_estimators.R
\name{meanipw}
\alias{meanipw}
\title{computes an ipw estimator of the mean of a random variable Y}
\usage{
meanipw(y, ps = NULL, x = NULL, type = "mean2")
}
\arguments{
\item{y}{A vector of outcomes (may contain NAs)}

\item{ps}{The estimated propensity score, i.e. the probability of y being observed conditional to \code{x} (either \code{x} or \code{ps} muste be provided)}

\item{x}{A matrix of covariates without an intercept (either \code{x} or \code{ps} muste be provided)}

\item{type}{mean1 or mean2 corresponding to two ways of computing the ipw estimators, dividing by different constants, analogous to IPW1 and IPW2 in Lunceford and Davidian (2008) respectively}
}
\value{
A list with elements pred (the propensity score), weights (the weight received by each observation) and mean (the estimated mean of Y)
}
\description{
computes an ipw estimator of the mean of a random variable Y
}
\examples{
n <- 100
alpha <- c(0, 0.1, -1.1)
beta <- c(0, -3, 10)
vars <- rsamplemissingnorm(n, alpha, beta)
x <- vars$x
y <- vars$y
a <- vars$a
meanipw(y, x = x)$mean
}
