% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_basic_functions.R
\name{pax}
\alias{pax}
\title{computes probability of a=1 conditional to x using the logistic model with parameter alpha}
\usage{
pax(x, alpha)
}
\arguments{
\item{x}{A matrix of covariates without the intercept}

\item{alpha}{A vector of parameters}
}
\value{
a vector representing the probability of a=1 conditional to x using the logistic model with parameter alpha (with intercept automatically added)
}
\description{
computes probability of a=1 conditional to x using the logistic model with parameter alpha
}
