% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_basic_functions.R
\name{pigorrospps}
\alias{pigorrospps}
\title{computes the propensity score using the spps model introduced in Molina et al 2018}
\usage{
pigorrospps(x, a)
}
\arguments{
\item{x}{A matrix of covariates without an intercept}

\item{a}{A vector of 0 and 1; if the i-th entry of a equals 1, it indicates the i-th outcome is missing.}
}
\value{
A vector indicating the estimated probabilities that a equals 1 conditional to x.
}
\description{
computes the propensity score using the spps model introduced in Molina et al 2018
}
