% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_estimators.R
\name{r}
\alias{r}
\title{Computes a regression estimator of the mean of a random variable Y based on least squares}
\usage{
r(x, y)
}
\arguments{
\item{x}{A matrix of covariates without an intercept}

\item{y}{A vector of outcomes (may contain NAs)}
}
\value{
The estimated mean of Y
}
\description{
Computes a regression estimator of the mean of a random variable Y based on least squares
}
\examples{
  vars <- rsamplemissing(n, alpha, beta)
  x <- vars$x
  y <- vars$y
  r(x, y)
}
