% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RamerDouglasPeucker}
\alias{RamerDouglasPeucker}
\title{Ramer-Douglas-Peucker}
\usage{
RamerDouglasPeucker(x, y, epsilon)
}
\arguments{
\item{x}{The \code{x} values of the curve.}

\item{y}{The \code{y} values of the curve.}

\item{epsilon}{The threshold for filtering outliers from the simplified curve.}
}
\value{
The \code{x} and \code{y} values of the simplified curve.
}
\description{
The Ramer-Douglas-Peucker algorithm for reducing the number of points on a curve.
}
