% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDP-package.R
\docType{package}
\name{RDP-package}
\alias{RDP}
\alias{RDP-package}
\title{\code{RDP}  package}
\description{
Implementation of the Ramer-Douglas-Peucker algorithm.
}
\author{
\strong{Maintainer}: Robert Dahl Jacobsen \email{git@dahl-jacobsen.dk}

}
