% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RDP}
\alias{RDP}
\title{Ramer-Douglas-Peucker}
\usage{
RDP(x, y, epsilon)
}
\arguments{
\item{x}{The \code{x} values}

\item{y}{The \code{y} values}

\item{epsilon}{The}
}
\value{
The \code{x} and \code{y} values.
}
\description{
The Ramer-Douglas-Peucker algorithm for reducing the number of points on a curve.
}
