% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubly_protected_estimators.R
\name{rdpn}
\alias{rdpn}
\title{Computes a robust doubly protected estimator of a quantile of a random variable Y with the robust method proposed in Sued, Valdora and Yohai (2019)}
\usage{
rdpn(x, y, ps = NULL, type = 2, qq = 0.5)
}
\arguments{
\item{x}{A matrix of covariates without an intercept (either \code{x} or \code{ps} muste be provided)}

\item{y}{A vector of outcomes (may contain NAs)}

\item{ps}{The estimated propensity score, i.e. the probability of y being observed conditional to \code{x} (either \code{x} or \code{ps} muste be provided)}

\item{type}{1 or 2; if type=2 the estimator is normalized as explained in Sued, Valdora and Yohai (2019)}
}
\value{
The estimated quantile of Y
}
\description{
Computes a robust doubly protected estimator of a quantile of a random variable Y with the robust method proposed in Sued, Valdora and Yohai (2019)
}
\examples{
  vars <- rsamplemissing(n, alpha, beta)
  x <- vars$x
  y <- vars$y
  rdpn(x, y)
}
