% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubly_protected_estimators.R
\name{rdpz}
\alias{rdpz}
\title{Computes a robust doubly protected estimator of a quantile of a random variable Y with the method proposed in Zhang et al. (2011) robustified as explained in Sued, Valdora and Yohai (2019)}
\usage{
rdpz(x, y, ps = NULL, qq = 0.5)
}
\arguments{
\item{x}{A matrix of covariates without an intercept (either \code{x} or \code{ps} muste be provided)}

\item{y}{A vector of outcomes (may contain NAs)}

\item{ps}{The estimated propensity score, i.e. the probability of y being observed conditional to \code{x} (either \code{x} or \code{ps} muste be provided)}
}
\value{
The estimated quantile of Y
}
\description{
Computes a robust doubly protected estimator of a quantile of a random variable Y with the method proposed in Zhang et al. (2011) robustified as explained in Sued, Valdora and Yohai (2019)
}
\examples{
  vars <- rsamplemissing(n, alpha, beta)
  x <- vars$x
  y <- vars$y
  pred <- pigorro(x, a)
  rdpz(x, y) #is the same as
  rdpz(x, y, ps = pred)
}
