% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_estimators.R
\name{rrz}
\alias{rrz}
\title{Computes a regression estimator of a quantile of a random variable Y with the robust method based on the method proposed Zhang et al (2011) described in Sued, Valdora and Yohai (2019)}
\usage{
rrz(x, y, ps = NULL, qq = 0.5)
}
\arguments{
\item{x}{A matrix of covariates without an intercept}

\item{y}{A vector of outcomes (may contain NAs)}
}
\value{
The estimated quantile of Y
}
\description{
Computes a regression estimator of a quantile of a random variable Y with the robust method based on the method proposed Zhang et al (2011) described in Sued, Valdora and Yohai (2019)
}
\examples{
  vars <- rsamplemissing(n, alpha, beta)
  x <- vars$x
  y <- vars$y
  a <- vars$a
  rrz(x, y)
}
