% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_generating.R
\name{rsamplemissingcauchy}
\alias{rsamplemissingcauchy}
\title{generates a sample with misssing values (x,a,y) that follows the distribution described in section  5 with errors following a cauchy distribution}
\usage{
rsamplemissingcauchy(n, alpha, beta)
}
\arguments{
\item{n}{The sample size}

\item{alpha}{A vector of parameters corresponding to the logsitic model that relates \code{a} and \code{x}}

\item{beta}{A vector of parameters corresponding to the linear model that relates \code{y} and \code{x}}
}
\value{
A list with elements x (the matrix of covariates), y (outcomes) and a (the missingness indicator)
}
\description{
generates a sample with misssing values (x,a,y) that follows the distribution described in section  5 with errors following a cauchy distribution
}
