% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_generating.R
\name{rsamplemissingt}
\alias{rsamplemissingt}
\title{generates a sample with misssing values (x,a,y) that follows the distribution described in section  5 with errors following a t distribution}
\usage{
rsamplemissingt(n, alpha, beta, df)
}
\arguments{
\item{n}{The sample size}

\item{alpha}{A vector of parameters corresponding to the logsitic model that relates \code{a} and \code{x}}

\item{beta}{A vector of parameters corresponding to the linear model that relates \code{y} and \code{x}}

\item{df}{The degrees of freedom of the t distribution of the errors}
}
\value{
A list with elements x (the matrix of covariates), y (outcomes) and a (the missingness indicator)
}
\description{
generates a sample with misssing values (x,a,y) that follows the distribution described in section  5 with errors following a t distribution
}
\examples{
n <- 100
alpha <- c(0, 0.1, -1.1)
beta <- c(0, -3, 10)
vars <- rsamplemissingnorm(n, alpha, beta)
}
