% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_estimators.R
\name{rz}
\alias{rz}
\title{Computes a regression estimator of a quantile of a random variable Y with the method proposed in Zhang et al (2011)}
\usage{
rz(x, y, ps = NULL, qq = 0.5)
}
\arguments{
\item{x}{A matrix of covariates without an intercept}

\item{y}{A vector of outcomes (may contain NAs)}
}
\value{
The estimated quantile of Y
}
\description{
Computes a regression estimator of a quantile of a random variable Y with the method proposed in Zhang et al (2011)
}
\examples{
  vars <- rsamplemissing(n, alpha, beta)
  x <- vars$x
  y <- vars$y
  a <- vars$a
  rz(x, y)
}
