% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_estimators.R
\name{sy}
\alias{sy}
\title{Computes a regression estimator of a quantile of a random variable Y with the robust method proposed in Sued and Yohai (2011)}
\usage{
sy(x, y, qq = 0.5)
}
\arguments{
\item{x}{A matrix of covariates without an intercept
#' @examples 
  vars <- rsamplemissing(n, alpha, beta)
  x <- vars$x
  y <- vars$y
  a <- vars$a
  sy(x, y)$quantile}

\item{y}{A vector of outcomes (may contain NAs)}
}
\value{
A list; element quantile is the estimated quantile of Y
}
\description{
Computes a regression estimator of a quantile of a random variable Y with the robust method proposed in Sued and Yohai (2011)
}
