% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDP-package.R
\docType{package}
\name{RDP-package}
\alias{RDP}
\alias{RDP-package}
\title{\code{RDP}  package}
\description{
Implementation of the \href{https://en.wikipedia.org/wiki/Ramer-Douglas-Peucker_algorithm}{Ramer-Douglas-Peucker algorithm}.
}
\references{
Urs Ramer (1972), "An iterative procedure for the polygonal approximation of plane curves".
\emph{Computer Graphics and Image Processing} \strong{1}, 244--256.
\doi{10.1016/S0146-664X(72)80017-0}.

David H. Douglas and Thomas K. Peucker (1973), "Algorithms for the Reduction of the Number of Points Required to Represent a Digitized Line or its Caricature".
\emph{Cartographica} \strong{10}, 112--122.
\doi{10.3138/FM57-6770-U75U-7727}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robertdj/RDP}
}

}
\author{
\strong{Maintainer}: Robert Dahl Jacobsen \email{cran@dahl-jacobsen.dk}

}
