% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDP-package.R
\docType{package}
\name{RDP-package}
\alias{RDP}
\alias{RDP-package}
\title{\code{RDP}  package}
\description{
Implementation of the \href{https://en.wikipedia.org/wiki/Ramer-Douglas-Peucker_algorithm}{Ramer-Douglas-Peucker algorithm}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robertdj/RDP}
}

}
\author{
\strong{Maintainer}: Robert Dahl Jacobsen \email{cran@dahl-jacobsen.dk}

}
